fixed send email function
