+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "PropertyPro-lite",
+  "name": "PropertyProLite",
   "version": "1.0.0",
   "description": "Property Pro Lite is a platform where people can create and/or search properties for sale or rent.",
   "main": "index.js",
@@ -33,6 +33,7 @@
     "express-validator": "^5.3.1",
     "moment": "^2.24.0",
     "nodemailer": "^6.2.1",
+    "sinon": "^7.3.2",
     "swagger-jsdoc": "^3.2.9",
     "uuid": "^3.3.2"
   },
